Was using ints in a bunch of places I shoulda been using schlongs. ... longs.

+++ src/main/java/au/com/samcday/jnntp/GroupInfo.java
@@ -1,7 +1,7 @@
 package au.com.samcday.jnntp;
 
 public class GroupInfo {
-    public int low;
-    public int high;
-    public int count;
+    public long low;
+    public long high;
+    public long count;
 }
